refactor: move out of date encounter routing project to archive
