+++ archive/EncounterTrackerDEFUNCT/scrapers/googleSheetScraper.py
@@ -1,0 +1,62 @@
+from __future__ import print_function
+
+import os.path
+
+from google.auth.transport.requests import Request
+from google.oauth2.credentials import Credentials
+from google_auth_oauthlib.flow import InstalledAppFlow
+from googleapiclient.discovery import build
+from googleapiclient.errors import HttpError
+
+import pandas as pd
+
+
+# If modifying these scopes, delete the file token.json.
+SCOPES = ['https://www.googleapis.com/auth/spreadsheets.readonly']
+
+def main(spreadsheetID, sampleRange):
+    """Shows basic usage of the Sheets API.
+    Prints values from a sample spreadsheet.
+    """
+    creds = None
+    # The file token.json stores the user's access and refresh tokens, and is
+    # created automatically when the authorization flow completes for the first
+    # time.
+    if os.path.exists('token.json'):
+        creds = Credentials.from_authorized_user_file('token.json', SCOPES)
+    # If there are no (valid) credentials available, let the user log in.
+    if not creds or not creds.valid:
+        if creds and creds.expired and creds.refresh_token:
+            creds.refresh(Request())
+        else:
+            flow = InstalledAppFlow.from_client_secrets_file(
+                'credentials.json', SCOPES)
+            creds = flow.run_local_server(port=0)
+        # Save the credentials for the next run
+        with open('token.json', 'w') as token:
+            token.write(creds.to_json())
+
+    try:
+        service = build('sheets', 'v4', credentials=creds)
+
+        # Call the Sheets API
+        sheet = service.spreadsheets()
+        result = sheet.values().get(spreadsheetId=spreadsheetID,
+                                    range=sampleRange).execute()
+        values = result.get('values', [])
+
+        if not values:
+            print('No data found.')
+            return
+        
+        return(values)
+
+
+    except HttpError as err:
+        print(err)
+
+# if __name__ == "__main__":
+#     spreadsheetID = "1nl-oAtNK19cbusswH7r3s_5Q-BfTwDPQIIDiTb-dJJc"
+#     sampleRange = 'Groups'
+
+#     main(spreadsheetID, sampleRange)
